examples/libraries/interactions_bot/requirements.txt
botstrap==0.2.3
discord-py-interactions==4.3.3
